Added Override start position and Equidistant Points Features
The start positions/rotations can be overridden by checking Booleans and changing the co-ordinates.

Added a feature where the rotation points are changed to a number of equidistant points if int variables are above 0 based on an end point -- i.e. if the endpoint was 180,  an int of 2 would convert to {0, 90, 180} since 2 transitions.