Update GaugeIndicator.cs (Full Script)
For the correctpoint boolean, it seemed to work coincidentally / in an obfuscated way. I think I corrected the logic.

The rotation points didn't work properly when going back and forth. I think this was a mistake from refactoring, because I knew straight away what the problem was. Did some brief testing and didn't notice anything strange.